# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Copilot Dashboard",
    "description": "Monitor all your GitHub Copilot CLI sessions in real-time.",
    "version": "0.4.1"
  },
  "paths": {
...
    "/api/session/{session_id}": {
      "get": {
        "summary": "Api Session Detail",
        "description": "Get detailed info for a single session.",
        "operationId": "api_session_detail_api_session__session_id__get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/kill/{session_id}": {
      "post": {
        "summary": "Api Kill",
        "description": "Kill the process for a running session.",
        "operationId": "api_kill_api_kill__session_id__post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/focus/{session_id}": {
      "post": {
        "summary": "Api Focus",
        "description": "Focus the terminal window for a running session.",
        "operationId": "api_focus_api_focus__session_id__post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActionResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "ActionResponse",
        "description": "Generic success/failure response for POST actions."
      },
      "CheckpointResponse": {
        "properties": {
          "checkpoint_number": {
            "type": "integer",
            "title": "Checkpoint Number"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "overview": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overview"
          },
          "next_steps": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Steps"
          }
        },
        "type": "object",
        "required": [
          "checkpoint_number"
        ],
        "title": "CheckpointResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RefResponse": {
        "properties": {
          "ref_type": {
            "type": "string",
            "title": "Ref Type"
          },
          "ref_value": {
            "type": "string",
            "title": "Ref Value"
          }
        },
        "type": "object",
        "required": [
          "ref_type",
          "ref_value"
        ],
        "title": "RefResponse"
      },
...
      "SessionDetailResponse": {
        "properties": {
          "checkpoints": {
            "items": {
              "$ref": "#/components/schemas/CheckpointResponse"
            },
            "type": "array",
            "title": "Checkpoints"
          },
          "refs": {
            "items": {
              "$ref": "#/components/schemas/RefResponse"
            },
            "type": "array",
            "title": "Refs"
          },
          "turns": {
            "items": {
              "$ref": "#/components/schemas/TurnResponse"
            },
            "type": "array",
            "title": "Turns"
          },
          "recent_output": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Recent Output"
          },
          "tool_counts": {
            "items": {
              "$ref": "#/components/schemas/ToolCountResponse"
            },
            "type": "array",
            "title": "Tool Counts"
          },
          "files": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Files"
          }
        },
        "type": "object",
        "title": "SessionDetailResponse"
      },
...
      "ToolCountResponse": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "name",
          "count"
        ],
        "title": "ToolCountResponse"
      },
      "TurnResponse": {
        "properties": {
          "turn_index": {
            "type": "integer",
            "title": "Turn Index"
          },
          "user_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Message"
          },
          "assistant_response": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assistant Response"
          }
        },
        "type": "object",
        "required": [
          "turn_index"
        ],
        "title": "TurnResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}